Add simple example of using array views with the SPIKE solver.
To illustrate the use of array views, this example program solves the following problem:
[A 0] [x1] = [b1]
[0 A] [x2]     [b2]
where the matrix A is read from a matrix market file.
b1 and b2 are views of the first and second half, respectively, of some array b of appropriate
size; similarly, x1 and x2 are views in a larger solution array x.

Usage:
    driver_views  FILENAME NUMPART