Refactor mixed-precision implementation.
We allow for a different precision in the preconditioner stage of the SPIKE solver
through an additional template parameter, named 'PrecValueType'.  All calculations
within the preconditioner are performed with this type and with containers based on it.

Solution is calculated with the precision implicitly defined by the 'Array' template
parameter used to instantiate the SPIKE solver.